feat(logs): Exposing logs options to public settings
Restructured the way that logs operate.
The `init(level)` function no longer receives a `level`. Instead, the setup
reads the settings from package.load. This allows to have the same behavior for
all the modules that require to use logs.

The `plenary` dependency is handled by a pcall statement. If the user hasn't
installed `plenary`, be notified and use the built-in logs. This also applies to
the logging level "none".