fix: Correctly differentiate between absolute and relative paths
Fixes the issue https://github.com/barreiroleo/ltex_extra.nvim/issues/41

Now, the behavior of missing subdirectories is the same as if the relative path were provided:

    If no subdirectory is missing, no directory is created.
    If only the last subdirectory is missing, it's created.
    If more than the one subdirectories are missing, the error is thrown.

Before this commit, if the relative path would be provided, and the last directory was missing,
it would be created. But, when the absolute path would be provided and the last directory was
missing, the absolute path would be treated as relative path. So the plugin would concatenate
root_dir and the absolute path, which would result in one long path, meaning not only the last
directory would be missing, but the whole second part of newly created path, and an error would
be thrown.