feat: Internal representation of ltex settings
This is an attempt to reduce the i/o operations by syncing this against
the server instead of be a proxy between disk and server.
Hopefully after that, ltex_extra will write on disk for exports and
reads just once at plugin start.

PS: Yeah, I'm not considering the fs watch in this hypothesis but I'd
say that edit these files is more likely an exception than a normal flow.